+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - select between `alphabetical` and `original` ordering for the feature list.
   - show two columns in the feature list if the feature has a second identifier associated.
 - Added support for generating view-config for Anndata-Zarr files that don't have .zmetadata file in the folder.
+- Turn on unit test coverage in Vitest config.
 
 ### Changed
 - Fix Material UI import statement.